clean up code, add documentation, delete the old pfilter_complete and rename the file.
